SVN_SILENT warning--
svn path=/branches/koffice/1.6/koffice/; revision=566376

+++ kexi/kexidb/queryschema.cpp
@@ -386,7 +386,7 @@ FieldList& QuerySchema::insertField(uint position, Field *field,
 			<<"' must contain table information!" <<endl;
 		return *this;
 	}
-	if (fieldCount()>=d->visibility.size()) {
+	if (fieldCount()>=(int)d->visibility.size()) {
 		d->visibility.resize(d->visibility.size()*2);
 		d->tablesBoundToColumns.resize(d->tablesBoundToColumns.size()*2);
 	}
@@ -443,7 +443,7 @@ FieldList& QuerySchema::insertField(uint position, Field *field,
 
 int QuerySchema::tableBoundToColumn(uint columnPosition) const
 {
-	if (columnPosition > d->tablesBoundToColumns.count()) {
+	if (columnPosition > (int)d->tablesBoundToColumns.count()) {
 		KexiDBWarn << "QuerySchema::tableBoundToColumn(): columnPosition (" << columnPosition
 			<< ") out of range" << endl;
 		return -1;
